create file

+++ 0x15-file_io/1-create_file.c
@@ -9,8 +9,8 @@
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	unsigned int i = 0;
-	ssize_t fd_w;
+	int i = 0;
+	int fd_w;
 
 	if (filename == NULL)
 	{
